requirements.txt
streamlit>=1.30.0
opencv-python>=4.8.0
pillow>=10.0.0
pyzbar>=0.1.9
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
responses>=0.24.0
